pyproject.toml
[project]
name = "gcpeasy"
version = "0.1.0"
description = "GCP, but easy peasy"
readme = "README.md"
authors = [
    { name = "mikonapoli", email = "michelangelo.bucci@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-cloud-bigquery",
    "pandas",
    "db-dtypes",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
